contrib: move shared Trace function to internal package
