+++ contrib/julienschmidt/httprouter/httprouter.go
@@ -9,7 +9,7 @@ import (
 	"github.com/DataDog/dd-trace-go/tracer"
 	"github.com/DataDog/dd-trace-go/tracer/ext"
 
-	httptrace "github.com/DataDog/dd-trace-go/contrib/net/http"
+	"github.com/DataDog/dd-trace-go/contrib/internal"
 )
 
 // Router is a traced version of httprouter.Router.
@@ -41,5 +41,5 @@ func (r *Router) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	resource := req.Method + " " + route
 
 	// we need to wrap the ServeHTTP method to be able to trace it
-	httptrace.Trace(r.Router, w, req, r.service, resource, r.Tracer)
+	internal.Trace(r.Router, w, req, r.service, resource, r.Tracer)
 }
